+++ src/views/v2/list.php
@@ -10,10 +10,10 @@
  * @link    http://evnt.is/1aiy
  *
  * @since 6.1.4 Changing our nonce verification structures.
+ * @since 6.2.0 Moved the header information into a new components/header.php template.
  * @since 6.12.0 Add aria-label to events list for improved accessibility.
  *
- * @version 6.2.0
- * @since 6.2.0 Moved the header information into a new components/header.php template.
+ * @version 6.12.0
  *
  * @var array    $events               The array containing the events.
  * @var string   $rest_url             The REST URL.