+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### [6.15.3] 2025-09-10
+
+* Tweak - * Tweak - Changed views: `blocks/event-venue`, `blocks/parts/map`, `blocks/parts/venue`, `modules/meta/map`, `modules/meta/organizer`, `modules/meta/venue`, `v2/components/header`, `v2/day`, `v2/day/event/venue`, `v2/list`, `v2/list/event/venue`
+* Security - Ensures password protected Venues and Organizers have their data protected. [SVUL-19]
+* Language - 0 new strings added, 59 updated, 0 fuzzied, and 0 obsoleted.
+
 ### [6.15.2] 2025-09-09
 
 * Fix - Corrected path for enqueuing ACF scripts to avoid 404 errors. [TEC-5604]